Add a CLI for use by netlify dev
