+++ package.json
@@ -3,6 +3,9 @@
   "version": "1.8.0",
   "description": "Netlify Build plugin to bundle Edge functions",
   "main": "src/index.js",
+  "bin": {
+    "plugin-build-edge-handlers": "src/cli.js"
+  },
   "scripts": {
     "test": "npm run format && npm run test:dev",
     "format": "run-s format:*",
