+++ test/helpers/run.js
@@ -1,3 +1,5 @@
+const childProcess = require("child_process");
+const path = require("path");
 const { env } = require("process");
 
 // Ensure @netlify/build logs have colors. Must be done before requiring it.
@@ -20,6 +22,29 @@ const runNetlifyBuild = async function (t, fixtureName, { expectedSuccess = true
   return { output };
 };
 
+/**
+ * Run a CLI build.
+ *
+ * @param {*} t the test function
+ * @param {*} fixtureName the fixture to bundle
+ */
+const runCliBuild = async (fixtureName, subdir) => {
+  const cliPath = path.join(path.dirname(path.dirname(__dirname)), "src/cli.js");
+  const fixturePath = path.join(path.dirname(__dirname), "fixtures", fixtureName, subdir || "edge-handlers");
+
+  const output = await new Promise((res, rej) => {
+    childProcess.execFile(cliPath, ["build", fixturePath], (err, stdout) => {
+      if (err) {
+        rej(err);
+      } else {
+        res(stdout);
+      }
+    });
+  });
+
+  return JSON.parse(output);
+};
+
 // @netlify/build stdout/stderr logs are concatenated as an array.
 // This flattens them as a string similar to what would be printed in real logs.
 const serializeOutput = function (logs) {
@@ -39,4 +64,4 @@ const isDebugMode = function () {
   return env.DEBUG === "1";
 };
 
-module.exports = { runNetlifyBuild };
+module.exports = { runCliBuild, runNetlifyBuild };
